TOPO: move topo to framework
