+++ .github/workflows/codestyle.yaml
@@ -36,7 +36,7 @@ jobs:
               return 1
             fi
           fi
-          if ! echo $msg | grep -qP '^Merge |^((CODESTYLE|REVIEW|CORE|UTIL|TEST|API|DOCS|TOOLS|BUILD|MC|SCHEDULE)|(CL/|TL/|MC/|UCP|UCG|NCCL|SHARP|BASIC|HIER|CUDA|CPU|EE))+: \w'
+          if ! echo $msg | grep -qP '^Merge |^((CODESTYLE|REVIEW|CORE|UTIL|TEST|API|DOCS|TOOLS|BUILD|MC|SCHEDULE|TOPO)|(CL/|TL/|MC/|UCP|UCG|NCCL|SHARP|BASIC|HIER|CUDA|CPU|EE))+: \w'
           then
             echo "Wrong header"
             return 1
